+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
-## v1.0.6 (18/04/2023)
+## v1.0.9 (05/05/2023)
+
+### Bugfix
+- Fixed setup.cfg bad imports
+
+## v1.0.7 (05/05/2023)
 
 ### Bugfix
 - Fixed imports
-
+- Added missing sql files to package
 ## v1.0.5 (17/04/2023)
 
 ### Feature